phe/py/requirements.txt
lightphe
web3
python-dotenv
